backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
openai>=1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pydantic==1.10.7
python-dotenv==1.0.0
pdfplumber>=0.9.0
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0  # For async HTTP requests 